% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleNeonorm-package.R
\docType{package}
\name{nimbleNeonorm-package}
\alias{nimbleNeonorm}
\alias{nimbleNeonorm-package}
\title{nimbleNeonorm: Neo-Normal Distributions for 'NIMBLE'}
\description{
Provides two flexible neo-normal probability distributions, MSNBurr and MSNBurr-IIa, with integration for the 'NIMBLE' Bayesian inference framework. Both distributions extend the normal distribution with additional shape parameters controlling skewness. The package provides density, distribution, quantile, and random generation functions for use in 'NIMBLE' models, with derivative support for Hamiltonian Monte Carlo (HMC) and No-U-Turn Sampler (NUTS) methods.
}
\author{
\strong{Maintainer}: Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})

Authors:
\itemize{
  \item Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})
  \item La Ode Rafiq Zandik Badarun
}

}
\keyword{internal}
