% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleNeonorm-package.R
\docType{package}
\name{nimbleNeonorm-package}
\alias{nimbleNeonorm}
\alias{nimbleNeonorm-package}
\title{nimbleNeonorm: Neo-Normal Distributions for 'NIMBLE'}
\description{
Provides the MSNBurr and MSNBurr-IIa neo-normal distributions as
user-defined distributions for 'NIMBLE', together with helpers that follow
the Bayesian workflow of Gelman et al. (2020).
}
\section{Workflow}{

\enumerate{
\item \strong{Distributions}: \link{MSNBurr}, \link{MSNBurr2a}; make them available in
model code with \code{\link[=register_neonorm]{register_neonorm()}}.
\item \strong{Prior predictive check}: \code{\link[=prior_predict_neonorm]{prior_predict_neonorm()}}.
\item \strong{Fitting} with several chains from dispersed starting values:
\code{\link[=runmcmc_neonorm]{runmcmc_neonorm()}}, \code{\link[=configure_mcmc_neonorm]{configure_mcmc_neonorm()}},
\code{\link[=prior_inits_neonorm]{prior_inits_neonorm()}}.
\item \strong{Convergence diagnostics} (\eqn{\hat{R}}, bulk and tail ESS):
\code{\link[=summarise_neonorm]{summarise_neonorm()}}.
\item \strong{Posterior predictive check}: \code{\link[=posterior_predict_neonorm]{posterior_predict_neonorm()}}.
\item \strong{Model comparison} with WAIC: \code{runmcmc_neonorm(..., WAIC = TRUE)}.
}

See \code{vignette("bayesian-workflow", package = "nimbleNeonorm")}.
}

\references{
Choir, A. S. (2020). \emph{Distribusi neo-normal baru dan karakteristiknya}
[Doctoral dissertation]. Institut Teknologi Sepuluh Nopember.

de Valpine, P., Turek, D., Paciorek, C. J., Anderson-Bergman, C.,
Temple Lang, D., & Bodik, R. (2017). Programming with models: Writing
statistical algorithms for general model structures with NIMBLE.
\emph{Journal of Computational and Graphical Statistics}, 26(2), 403--413.
\doi{10.1080/10618600.2016.1172487}

Gelman, A., Vehtari, A., Simpson, D., Margossian, C. C., Carpenter, B.,
Yao, Y., Kennedy, L., Gabry, J., Bürkner, P.-C., & Modrák, M. (2020).
\emph{Bayesian workflow}. arXiv:2011.01808.
\doi{10.48550/arXiv.2011.01808}

Ramadani, E. P., Choir, A. S., Pravitasari, A. A., & Paraguison, J.
(2025). Adding MSNBURR-IIa distribution to MultiBUGS. \emph{Jurnal Aplikasi
Statistika & Komputasi Statistik}, 17(2), 109--130.
\doi{10.34123/jurnalasks.v17i2.804}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafiqzand/nimbleNeonorm}
  \item Report bugs at \url{https://github.com/rafiqzand/nimbleNeonorm/issues}
}

}
\author{
\strong{Maintainer}: Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})

Authors:
\itemize{
  \item Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})
  \item La Ode Rafiq Zandik Badarun
}

}
\keyword{internal}
